core/requirements.txt
# general modules
Django>=3.2,<4.0
psycopg2-binary
dj-database-url
pillow
coreapi==2.3.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna==3.4
python-decouple==3.8

# third party modules
watchdog==2.1.3
requests==2.31.0
pyTelegramBotAPI==4.14.0
drf-yasg[validation]
django-grappelli==3.0.8
